cartcontroller changes

+++ TeamMaddog/Models/Product.cs
@@ -9,7 +9,7 @@ namespace TeamMaddog.Models
     {
         public int ProductID { get; set; }
         public string ProdName { get; set; }
-        public decimal ProdPrice { get; set; }
+        public double ProdPrice { get; set; }
         public string ProdSize { get; set; }
 
     }
